+++ packages/spacecat-shared-cloudflare-client/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [@adobe/spacecat-shared-cloudflare-client-v1.1.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-cloudflare-client-v1.1.0...@adobe/spacecat-shared-cloudflare-client-v1.1.1) (2026-06-26)
+
+### Bug Fixes
+
+* **cloudflare-client:** use GET for worker existence check before deploy ([#1736](https://github.com/adobe/spacecat-shared/issues/1736)) ([c9928d8](https://github.com/adobe/spacecat-shared/commit/c9928d8c03382cde7c5a01479ca1843c971200b7))
+
+## [@adobe/spacecat-shared-cloudflare-client-v1.1.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-cloudflare-client-v1.0.0...@adobe/spacecat-shared-cloudflare-client-v1.1.0) (2026-06-26)
+
+### Features
+
+* **cloudflare-client:** add zone account filter and overwrite guard on deploy ([#1728](https://github.com/adobe/spacecat-shared/issues/1728)) ([6f8cbe6](https://github.com/adobe/spacecat-shared/commit/6f8cbe6556123980b20890aacddf20a8033990cc))
+
 ## @adobe/spacecat-shared-cloudflare-client-v1.0.0 (2026-06-25)
 
 ### Features